librdmacm: Add support for epoll: epoll_wait
This commit implements epoll_wait to retrieve events processed by the
centralized thread for an epoll instance.

When epoll_wait is called, it copies the events collected by the global
thread from the ready list in the epoll instance to the user-provided
events buffer.

If no events are available in the `revents` field, the function triggers
the thread to recheck for events. Epoll_wait returns the total number of
ready events.

Signed-off-by: Batsheva Black <bblack@nvidia.com>